fix: disallow the same projectile to be shared by all turrets
Grenade types were implemented by the way of replacing the projectile reference of the turret def, but since the turret def stayed the same, changing a grenade type was changing it for all turrets.

Because of that, the chosen projectile was also reverting back to the default one on startup. The fix is replacing the whole turret def instead.